+++ packages/lumx-react/src/utils/date/isDateValid.ts
@@ -1,0 +1,4 @@
+/**
+ * Check if given date is valid.
+ */
+export const isDateValid = (date?: Date) => date instanceof Date && !Number.isNaN(date.getTime());