+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   `@lumx/react` no long depend on `moment` or `moment-range` to generate the date picker.
+-   Deprecated `@lumx/core/js/date-picker` functions that **will be removed in the next major version** along with `moment` and `moment-range`.
+-   DatePicker & DatePickerField: `locale` prop is now optional (uses browser locale by default)
+
 ## [3.5.3][] - 2023-08-30
 
 ### Changed
@@ -1807,8 +1813,6 @@ _Failed released_
 [3.5.0]: https://github.com/lumapps/design-system/tree/v3.5.0
 [unreleased]: https://github.com/lumapps/design-system/compare/v3.5.1...HEAD
 [3.5.1]: https://github.com/lumapps/design-system/tree/v3.5.1
-
-
-[Unreleased]: https://github.com/lumapps/design-system/compare/v3.5.3...HEAD
+[unreleased]: https://github.com/lumapps/design-system/compare/v3.5.3...HEAD
 [3.5.3]: https://github.com/lumapps/design-system/compare/v3.5.2...v3.5.3
 [3.5.2]: https://github.com/lumapps/design-system/tree/v3.5.2